+++ edi_party_data_oca/__init__.py
@@ -1,2 +1,1 @@
 from . import components
-from . import models
