+++ edi_party_data_oca/components/common.py
@@ -2,13 +2,19 @@
 # @author: Simone Orsi <simahawk@gmail.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo.tools import DotDict
-
 from odoo.addons.component.core import AbstractComponent
 
 
 class EDIExchangePartyDataMixin(AbstractComponent):
-    """Abstract component mixin provide partner data for exchanges."""
+    """Abstract component mixin provide partner data for exchanges.
+
+    .. deprecated::
+        Business logic lives in the `edi_party_helper_oca` module, which has
+        no component dependency: `EDIParty` (plain dataclass) or the
+        `edi.party.helper` abstract model. This component only wraps it for
+        backward compatibility and will be dropped once its usages are
+        ported over.
+    """
 
     _name = "edi.party.data.mixin"
     _inherit = "edi.component.mixin"
@@ -25,59 +31,20 @@ class EDIExchangePartyDataMixin(AbstractComponent):
         # based on current partner (eg: pick the parent).
         return self.work.partner
 
+    def _get_party(self):
+        return self.env["edi.party.helper"].get_party(
+            self.exchange_record,
+            self.partner,
+            name_field=getattr(self.work, "party_data_name_field", "name"),
+            lang_code=getattr(self.work, "lang", False),
+        )
+
     def get_party(self):
         """Return party information.
 
         Requires a res.partner to be passed via work context.
 
-        :return: odoo.tools.DotDict
+        :return: `EDIParty` (behaves like a read-only dict too, for
+            backward compatibility with the old ``DotDict``-based return)
         """
-        return self._party_from_partner()
-
-    def _party_from_partner(self, **kw):
-        # NB: for UBL this should probably replace
-        # `base.ubl._ubl_get_party_identification` which does nothing today.
-        party = DotDict(
-            name=self._get_name(),
-            identifiers=self._get_identifiers(),
-            endpoint=self._get_endpoint(),
-            lang=self._get_lang(),
-            partner=self.partner,
-        )
-        party.update(kw)
-        return party
-
-    def _get_name(self):
-        name_field = getattr(self.work, "party_data_name_field", "name")
-        return self.partner[name_field]
-
-    def _get_endpoint(self):
-        return {}
-
-    def _get_identifiers(self):
-        identifiers = self.partner.id_numbers.filtered(
-            lambda x: self._filter_id_number(x)
-        )
-        return [self._get_indentity(x) for x in identifiers]
-
-    def _filter_id_number(self, id_number):
-        if self.allowed_id_categories:
-            return id_number.category_id in self.allowed_id_categories
-        return True
-
-    def _get_indentity(self, id_number):
-        return DotDict(
-            attrs={
-                "schemeID": id_number.category_id.code,
-            },
-            value=id_number.name,
-        )
-
-    def _get_lang(self):
-        lang_code = getattr(self.work, "lang", False) or self.partner.lang
-        if not lang_code:
-            return False
-        lang = self.env["res.lang"]._get_data(code=lang_code)
-        if not lang:
-            return False
-        return {"name": lang.name, "code": lang.code, "short": lang.code.split("_")[0]}
+        return self._get_party()
