--- edi_party_data_oca/models/__init__.py
@@ -1,1 +1,0 @@
-from . import edi_exchange_type
