+++ edi_party_data_oca/readme/CONFIGURE.md
@@ -1,9 +1,6 @@
 ## ID numbers selection
 
-On the exchange type form, find the field "ID categories" and set the
-categories allowed for that exchange type.
-
-If not set, *all the IDs* of the partner will be exposed.
+See `edi_party_helper_oca`'s configuration.
 
 ## Name field
