+++ edi_party_data_oca/tests/test_party_data.py
@@ -2,91 +2,17 @@
 # @author: Simone Orsi <simahawk@gmail.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo.addons.edi_component_oca.tests.common import EDIBackendCommonComponentTestCase
-
 from ..utils import get_party_data_component
+from .common import PartyDataCommonTestCase
+
+# Business logic is tested against the component-independent helper in
+# `test_party_helper.py`. This suite only exercises the component wiring
+# (work context plumbing, hooks) on top of it.
 
 
-class PartyDataTestCase(EDIBackendCommonComponentTestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.backend = cls.env.ref("edi_core_oca.demo_edi_backend")
-        cls.cat_model = cls.env["res.partner.id_category"]
-        cls.all_cat = cls.cat_model.browse()
-        for i in range(1, 4):
-            rec = cls.cat_model.create({"code": f"cat{i}", "name": f"Cat {i}"})
-            cls.all_cat += rec
-            setattr(cls, f"category{i}", rec)
-
-        parent = cls.env["res.partner"].create(
-            {
-                "name": "ACME inc",
-                "is_company": True,
-            }
-        )
-
-        for i in range(1, 4):
-            rec = cls.env["res.partner"].create(
-                {
-                    "name": f"Test Partner {i}",
-                    "parent_id": parent.id,
-                    "id_numbers": [
-                        (
-                            0,
-                            0,
-                            {
-                                "name": f"{cat.code}-p{i}",
-                                "category_id": cat.id,
-                            },
-                        )
-                        for cat in cls.all_cat[i - 1 :]
-                    ],
-                }
-            )
-            setattr(cls, f"partner{i}", rec)
-
-        # No need for special file name gen
-        cls.exc_type = cls._create_exchange_type(
-            name="ID output test",
-            code="id_out_test",
-            direction="output",
-        )
-        cls.exc_record = cls.backend.create_record("id_out_test", {})
-
+class PartyDataTestCase(PartyDataCommonTestCase):
     def _get_provider(self, partner, **kw):
         return get_party_data_component(self.exc_record, partner, **kw)
-
-    def _expected_lang(self, partner):
-        if not partner.lang:
-            return False
-        lang = self.env["res.lang"]._get_data(code=partner.lang)
-        if not lang:
-            return False
-        return {"name": lang.name, "code": lang.code, "short": lang.code.split("_")[0]}
-
-    def _make_expected_data(
-        self, partner, number, allowed_codes=None, name_field="name", **kw
-    ):
-        data = {
-            "name": partner[name_field],
-            "identifiers": [
-                {"attrs": {"schemeID": "cat1"}, "value": f"cat1-p{number}"},
-                {"attrs": {"schemeID": "cat2"}, "value": f"cat2-p{number}"},
-                {"attrs": {"schemeID": "cat3"}, "value": f"cat3-p{number}"},
-            ],
-            "endpoint": {},
-            "lang": self._expected_lang(partner),
-            "partner": partner,
-        }
-        data.update(kw)
-        if allowed_codes:
-            data["identifiers"] = [
-                x
-                for x in data["identifiers"]
-                if x["attrs"]["schemeID"] in allowed_codes
-            ]
-        return data
 
     def test_lookup(self):
         provider = self._get_provider(self.partner1)
@@ -94,69 +20,27 @@ class PartyDataTestCase(EDIBackendCommonComponentTestCase):
         self.assertFalse(provider.allowed_id_categories)
 
     def test_data(self):
-        expected = (
-            (self.partner1, self._make_expected_data(self.partner1, 1)),
-            (
-                self.partner2,
-                self._make_expected_data(
-                    self.partner2, 2, allowed_codes=["cat2", "cat3"]
-                ),
-            ),
-            (
-                self.partner3,
-                self._make_expected_data(self.partner3, 3, allowed_codes=["cat3"]),
-            ),
-        )
-        for partner, expected_data in expected:
-            provider = self._get_provider(partner)
-            res = provider.get_party()
-            self.assertEqual(res, expected_data)
-
-    def test_data_fullname_override(self):
-        # `name` is the default `party_data_name_field` since it doesn't embed
-        # multi-company/disambiguation suffixes the way `display_name` does;
-        # `display_name` is still available by explicit override.
-        expected = (
-            (
-                self.partner1,
-                self._make_expected_data(self.partner1, 1, name_field="display_name"),
-            ),
-            (
-                self.partner2,
-                self._make_expected_data(
-                    self.partner2,
-                    2,
-                    allowed_codes=["cat2", "cat3"],
-                    name_field="display_name",
-                ),
-            ),
-            (
-                self.partner3,
-                self._make_expected_data(
-                    self.partner3, 3, allowed_codes=["cat3"], name_field="display_name"
-                ),
-            ),
-        )
-        for partner, expected_data in expected:
-            provider = self._get_provider(
-                partner, work_ctx={"party_data_name_field": "display_name"}
-            )
-            res = provider.get_party()
-            self.assertEqual(res, expected_data)
-
-    def test_lang(self):
-        # No lang set on the partner -> no `lang` key rendered.
-        self.partner1.lang = False
+        expected = self._make_expected_data(self.partner1, 1)
         provider = self._get_provider(self.partner1)
-        self.assertFalse(provider.get_party()["lang"])
-        # Partner's own lang is used by default.
-        self.partner1.lang = "en_US"
-        provider = self._get_provider(self.partner1)
-        self.assertEqual(
-            provider.get_party()["lang"],
-            {"name": "English (US)", "code": "en_US", "short": "en"},
+        self.assertEqual(provider.get_party(), expected)
+
+    def test_data_limited(self):
+        self.exc_type.id_category_ids = self.category2
+        expected = self._make_expected_data(self.partner2, 2, allowed_codes=["cat2"])
+        provider = self._get_provider(self.partner2)
+        self.assertEqual(provider.get_party(), expected)
+
+    def test_data_name_field_override(self):
+        # `party_data_name_field` on the work context is forwarded to the
+        # party's `name_field`.
+        expected = self._make_expected_data(self.partner1, 1, name_field="display_name")
+        provider = self._get_provider(
+            self.partner1, work_ctx={"party_data_name_field": "display_name"}
         )
-        # `work_ctx["lang"]` takes precedence over the partner's own lang.
+        self.assertEqual(provider.get_party(), expected)
+
+    def test_lang_override_via_work_ctx(self):
+        # `lang` on the work context is forwarded to the party's `lang_code`.
         self.partner2.lang = False
         provider = self._get_provider(self.partner2, work_ctx={"lang": "en_US"})
         self.assertEqual(
@@ -167,57 +51,3 @@ class PartyDataTestCase(EDIBackendCommonComponentTestCase):
     def test_partner_in_party_data(self):
         provider = self._get_provider(self.partner1)
         self.assertEqual(provider.get_party()["partner"], self.partner1)
-
-    def test_data_limited_1(self):
-        self.exc_type.id_category_ids = self.category1
-        expected = (
-            (
-                self.partner1,
-                self._make_expected_data(self.partner1, 1, allowed_codes=["cat1"]),
-            ),
-            (self.partner2, self._make_expected_data(self.partner2, 2, identifiers=[])),
-            (self.partner3, self._make_expected_data(self.partner3, 3, identifiers=[])),
-        )
-        for partner, expected_data in expected:
-            provider = self._get_provider(partner)
-            res = provider.get_party()
-            self.assertEqual(res, expected_data)
-
-    def test_data_limited_2(self):
-        self.exc_type.id_category_ids = self.category2
-        expected = (
-            (
-                self.partner1,
-                self._make_expected_data(self.partner1, 1, allowed_codes=["cat2"]),
-            ),
-            (
-                self.partner2,
-                self._make_expected_data(self.partner2, 2, allowed_codes=["cat2"]),
-            ),
-            (self.partner3, self._make_expected_data(self.partner3, 3, identifiers=[])),
-        )
-        for partner, expected_data in expected:
-            provider = self._get_provider(partner)
-            res = provider.get_party()
-            self.assertEqual(res, expected_data)
-
-    def test_data_limited_3(self):
-        self.exc_type.id_category_ids = self.category3
-        expected = (
-            (
-                self.partner1,
-                self._make_expected_data(self.partner1, 1, allowed_codes=["cat3"]),
-            ),
-            (
-                self.partner2,
-                self._make_expected_data(self.partner2, 2, allowed_codes=["cat3"]),
-            ),
-            (
-                self.partner3,
-                self._make_expected_data(self.partner3, 3, allowed_codes=["cat3"]),
-            ),
-        )
-        for partner, expected_data in expected:
-            provider = self._get_provider(partner)
-            res = provider.get_party()
-            self.assertEqual(res, expected_data)
